Move Header & Footer
This also flattened the dir structure of contained subcomponents.
